requirements.txt
numpy~=1.22.2
pip~=21.3.1
wheel~=0.37.1
MarkupSafe~=2.1.0
Werkzeug~=2.0.3
Jinja2~=3.0.3
click~=8.0.4
itsdangerous~=2.1.0
setuptools~=60.6.0
Flask~=2.0.3
Pillow~=9.0.1
rsa~=4.8
pyasn1~=0.4.8
six~=1.16.0
gast~=0.5.3
h5py~=3.6.0
keras~=2.8.0
tensorflow~=2.8.0
scipy~=1.8.0
wrapt~=1.14.0
requests~=2.27.1
cachetools~=5.0.0
certifi~=2021.10.8
urllib3~=1.26.8
Markdown~=3.3.6
oauthlib~=3.2.0
idna~=3.3
pyparsing~=3.0.7
tensorboard~=2.8.0
flatbuffers~=2.0
astunparse~=1.6.3
termcolor~=1.1.0
zipp~=3.7.0
joblib~=1.1.0
scikit-learn~=1.0.2
threadpoolctl~=3.1.0